refactor(keyboard): group keycap item factory files
